move all components files to a folder
