+++ src/App.js
@@ -3,9 +3,9 @@ import React, { Component } from 'react';
 import logoyellowbg from './logoyellowbg.svg';
 
 import './App.css';
-import LeftNav from "./LeftNav"
-import Feed from "./Feed"
-import RightNav from "./RightNav"
+import LeftNav from "./components/LeftNav"
+import Feed from "./components/Feed"
+import RightNav from "./components/RightNav"
 
 
 class App extends Component {
