+++ src/components/NewPost.js
